Revert "Add alienc0der's v0.0.7 changes"
