+++ .github/workflows/syrius_builder.yml
@@ -13,6 +13,8 @@ env:
 
 jobs:
   build-macos:
+    env:
+        WALLET_CONNECT_PROJECT_ID: ${{ secrets.WC_PROJECT_ID }}
     runs-on: macos-12
     steps:
       - name: Checkout
@@ -33,7 +35,7 @@ jobs:
       - name: Build syrius desktop 
         run: |
           flutter config --enable-macos-desktop
-          flutter build macos --release
+          flutter build macos --dart-define=WC_PROJECT_ID=$WALLET_CONNECT_PROJECT_ID --release
       - name: Package into DMG
         run: |
           create-dmg --volname syrius \
@@ -55,6 +57,8 @@ jobs:
           name: macos-artifacts
           path: syrius-alphanet-macos-universal.dmg
   build-windows:
+    env:
+        WALLET_CONNECT_PROJECT_ID: ${{ secrets.WC_PROJECT_ID }}
     runs-on: windows-latest
     steps:
       - name: Checkout
@@ -69,7 +73,7 @@ jobs:
       - name: Build syrius desktop
         run: |
           flutter config --enable-windows-desktop
-          flutter build windows --release   
+          flutter build windows --dart-define=WC_PROJECT_ID=$env:WALLET_CONNECT_PROJECT_ID --release   
       - name: Package into zip
         run: |
           Compress-Archive -Path build\windows\runner\Release\* -DestinationPath .\syrius-alphanet-windows-amd64.zip
@@ -79,6 +83,8 @@ jobs:
           name: windows-artifacts
           path: syrius-alphanet-windows-amd64.zip
   build-linux:
+    env:
+        WALLET_CONNECT_PROJECT_ID: ${{ secrets.WC_PROJECT_ID }}
     runs-on: ubuntu-20.04
     steps:
       - name: Checkout
@@ -102,7 +108,7 @@ jobs:
       - name: Build syrius desktop
         run: |
           flutter config --enable-linux-desktop
-          flutter build linux --release -v
+          flutter build linux --dart-define=WC_PROJECT_ID=$WALLET_CONNECT_PROJECT_ID --release -v
       - name: Package zip
         run: |
           cd build/linux/x64/release/bundle
