+++ lib/blocs/auto_receive_tx_worker.dart
@@ -9,7 +9,6 @@ import 'package:zenon_syrius_wallet_flutter/main.dart';
 import 'package:zenon_syrius_wallet_flutter/model/model.dart';
 import 'package:zenon_syrius_wallet_flutter/utils/account_block_utils.dart';
 import 'package:zenon_syrius_wallet_flutter/utils/address_utils.dart';
-import 'package:zenon_syrius_wallet_flutter/utils/constants.dart';
 import 'package:zenon_syrius_wallet_flutter/utils/global.dart';
 import 'package:znn_sdk_dart/znn_sdk_dart.dart';
 
@@ -23,45 +22,7 @@ class AutoReceiveTxWorker extends BaseBloc<WalletNotification> {
     return _instance!;
   }
 
-  Future<void> autoReceiveTransactionHash(Hash currentHash) async {
-    if (!running) {
-      running = true;
-      try {
-        String toAddress =
-            (await zenon!.ledger.getAccountBlockByHash(currentHash))!
-                .toAddress
-                .toString();
-        KeyPair keyPair = kKeyStore!.getKeyPair(
-          kDefaultAddressList.indexOf(toAddress),
-        );
-        AccountBlockTemplate transactionParams = AccountBlockTemplate.receive(
-          currentHash,
-        );
-        AccountBlockTemplate response =
-            await AccountBlockUtils.createAccountBlock(
-          transactionParams,
-          'receive transaction',
-          blockSigningKey: keyPair,
-          waitForRequiredPlasma: true,
-        );
-        _sendSuccessNotification(response, toAddress);
-      } on RpcException catch (e, stackTrace) {
-        _sendErrorNotification(e.toString());
-        Logger('AutoReceiveTxWorker')
-            .log(Level.WARNING, 'autoReceive', e, stackTrace);
-      }
-      running = false;
-    }
-  }
-
   Future<void> autoReceive() async {
-    if (!sharedPrefsService!.get(
-      kAutoReceiveKey,
-      defaultValue: kAutoReceiveDefaultValue,
-    )) {
-      pool.clear();
-      return;
-    }
     // Make sure that AutoUnlockHtlcWorker is not running since it should be
     // given priority to send transactions.
     if (pool.isNotEmpty && !running && !sl<AutoUnlockHtlcWorker>().running) {
@@ -85,12 +46,8 @@ class AutoReceiveTxWorker extends BaseBloc<WalletNotification> {
           blockSigningKey: keyPair,
           waitForRequiredPlasma: true,
         );
+        pool.removeFirst();
         _sendSuccessNotification(response, toAddress);
-        if (pool.isNotEmpty) {
-          pool.removeFirst();
-        }
-        running = false;
-        autoReceive();
       } on RpcException catch (e, stackTrace) {
         _sendErrorNotification(e.toString());
         Logger('AutoReceiveTxWorker')
@@ -108,19 +65,6 @@ class AutoReceiveTxWorker extends BaseBloc<WalletNotification> {
       }
       running = false;
     }
-    return;
-  }
-
-  Future<void> addHash(Hash hash) async {
-    zenon!.stats.syncInfo().then((syncInfo) {
-      if (!pool.contains(hash) &&
-          (syncInfo.state == SyncState.syncDone ||
-              (syncInfo.targetHeight > 0 &&
-                  syncInfo.currentHeight > 0 &&
-                  (syncInfo.targetHeight - syncInfo.currentHeight) < 3))) {
-        pool.add(hash);
-      }
-    });
   }
 
   void _sendErrorNotification(String errorText) {
@@ -144,5 +88,21 @@ class AutoReceiveTxWorker extends BaseBloc<WalletNotification> {
         type: NotificationType.paymentReceived,
       ),
     );
+  }
+
+  Future<void> addHash(Hash hash) async {
+    if (!pool.contains(hash)) {
+      zenon!.stats.syncInfo().then((syncInfo) {
+        // Verify that the pool does not already contain the hash after the
+        // asynchronous request has completed and that the node is in sync.
+        if (!pool.contains(hash) &&
+            (syncInfo.state == SyncState.syncDone ||
+                (syncInfo.targetHeight > 0 &&
+                    syncInfo.currentHeight > 0 &&
+                    (syncInfo.targetHeight - syncInfo.currentHeight) < 3))) {
+          pool.add(hash);
+        }
+      });
+    }
   }
 }
