--- lib/blocs/wallet_connect/chains/i_chain.dart
@@ -1,5 +1,0 @@
-abstract class IChain {
-  String getNamespace();
-  String getChainId();
-  List<String> getEvents();
-}
