+++ lib/main.dart
@@ -18,16 +18,13 @@ import 'package:window_manager/window_manager.dart';
 import 'package:zenon_syrius_wallet_flutter/blocs/auto_unlock_htlc_worker.dart';
 import 'package:zenon_syrius_wallet_flutter/blocs/blocs.dart';
 import 'package:zenon_syrius_wallet_flutter/handlers/htlc_swaps_handler.dart';
-import 'package:zenon_syrius_wallet_flutter/blocs/wallet_connect/chains/i_chain.dart';
-import 'package:zenon_syrius_wallet_flutter/blocs/wallet_connect/chains/nom_service.dart';
 import 'package:zenon_syrius_wallet_flutter/blocs/wallet_connect/wallet_connect_pairings_bloc.dart';
 import 'package:zenon_syrius_wallet_flutter/blocs/wallet_connect/wallet_connect_sessions_bloc.dart';
 import 'package:zenon_syrius_wallet_flutter/model/model.dart';
 import 'package:zenon_syrius_wallet_flutter/screens/screens.dart';
 import 'package:zenon_syrius_wallet_flutter/services/htlc_swaps_service.dart';
-import 'package:zenon_syrius_wallet_flutter/services/i_web3wallet_service.dart';
 import 'package:zenon_syrius_wallet_flutter/services/shared_prefs_service.dart';
-import 'package:zenon_syrius_wallet_flutter/services/web3wallet_service.dart';
+import 'package:zenon_syrius_wallet_flutter/services/wallet_connect_service.dart';
 import 'package:zenon_syrius_wallet_flutter/utils/utils.dart';
 import 'package:zenon_syrius_wallet_flutter/widgets/widgets.dart';
 import 'package:znn_sdk_dart/znn_sdk_dart.dart';
@@ -37,9 +34,6 @@ SharedPrefsService? sharedPrefsService;
 HtlcSwapsService? htlcSwapsService;
 
 final sl = GetIt.instance;
-
-IWeb3WalletService? web3WalletService;
-final globalNavigatorKey = GlobalKey<NavigatorState>();
 
 main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -76,13 +70,8 @@ main() async {
   windowManager.ensureInitialized();
   await windowManager.setPreventClose(true);
 
-  web3WalletService = Web3WalletService();
-  web3WalletService!.create();
-
   // Setup services
   setup();
-
-  await web3WalletService!.init();
 
   // Setup local_notifier
   await localNotifier.setup(
@@ -178,13 +167,7 @@ void setup() {
       (() => SharedPrefsService.getInstance().then((value) => value!)));
   sl.registerSingleton<HtlcSwapsService>(HtlcSwapsService.getInstance());
 
-  // Initialize WalletConnect service
-  sl.registerSingleton<IWeb3WalletService>(web3WalletService!);
-  sl.registerSingleton<IChain>(
-    NoMService(reference: NoMChainId.mainnet),
-    instanceName: NoMChainId.mainnet.chain(),
-  );
-
+  sl.registerLazySingleton<WalletConnectService>(() => WalletConnectService());
   sl.registerSingleton<AutoReceiveTxWorker>(AutoReceiveTxWorker.getInstance());
   sl.registerSingleton<AutoUnlockHtlcWorker>(
       AutoUnlockHtlcWorker.getInstance());
@@ -296,7 +279,6 @@ class _MyAppState extends State<MyApp> with WindowListener, TrayListener {
                         child: Layout(
                           child: MaterialApp(
                             title: 's y r i u s',
-                            navigatorKey: globalNavigatorKey,
                             debugShowCheckedModeBanner: false,
                             theme: AppTheme.lightTheme,
                             darkTheme: AppTheme.darkTheme,
