+++ lib/utils/constants.dart
@@ -4,7 +4,8 @@ import 'package:zenon_syrius_wallet_flutter/widgets/widgets.dart';
 import 'package:znn_sdk_dart/znn_sdk_dart.dart';
 
 // WalletConnect
-const String kWcProjectId = '6106aa8c2f308b338f31465bef999a1f';
+const String kWcProjectId = String.fromEnvironment('WC_PROJECT_ID',
+    defaultValue: '6106aa8c2f308b338f31465bef999a1f');
 const String kZenonNameSpace = 'zenon';
 const walletConnectDirName = 'walletconnect';
 
@@ -108,12 +109,12 @@ const int kAmountInputMaxCharacterLength = 21;
 const int kSecondsPerMomentum = 10;
 const int kMaxP2pSwapsToStore = 500;
 
-const List<int> kNormalUsersPlasmaRequirements = [
+final List<int> kNormalUsersPlasmaRequirements = [
   kStakePlasmaAmountNeeded,
   kDelegatePlasmaAmountNeeded,
 ];
 
-const List<int> kPowerUsersPlasmaRequirements = [
+final List<int> kPowerUsersPlasmaRequirements = [
   kPillarPlasmaAmountNeeded,
   kSentinelPlasmaAmountNeeded,
   kIssueTokenPlasmaAmountNeeded,
@@ -121,6 +122,11 @@ const List<int> kPowerUsersPlasmaRequirements = [
 
 const String kLocalhostDefaultNodeUrl = 'ws://127.0.0.1:$kDefaultPort';
 const int kDefaultPort = 35998;
+
+List<String> kDefaultNodes = [
+  'Embedded Node',
+  kLocalhostDefaultNodeUrl,
+];
 
 const List<String> kWalletActions = [
   'pillar',
@@ -171,12 +177,9 @@ const String kAutoEraseNumAttemptsKey = 'auto_erase_num_attempts';
 const String kLaunchAtStartupKey = 'launch_at_startup';
 const String kEnableDesktopNotificationsKey = 'enable_desktop_notifications';
 const String kEnableClipboardWatcherKey = 'enable_clipboard_watcher';
-const String kAutoReceiveKey = 'enable_auto_receive';
-
 const bool kLaunchAtStartupDefaultValue = false;
 const bool kEnableDesktopNotificationsDefaultValue = false;
 const bool kEnableClipboardWatcherDefaultValue = false;
-const bool kAutoReceiveDefaultValue = true;
 
 /// Node management
 const String kChainIdKey = 'chain_id';
