+++ lib/utils/init_utils.dart
@@ -5,6 +5,7 @@ import 'package:hive/hive.dart';
 import 'package:zenon_syrius_wallet_flutter/handlers/htlc_swaps_handler.dart';
 import 'package:zenon_syrius_wallet_flutter/main.dart';
 import 'package:zenon_syrius_wallet_flutter/services/shared_prefs_service.dart';
+import 'package:zenon_syrius_wallet_flutter/services/wallet_connect_service.dart';
 import 'package:zenon_syrius_wallet_flutter/utils/address_utils.dart';
 import 'package:zenon_syrius_wallet_flutter/utils/constants.dart';
 import 'package:zenon_syrius_wallet_flutter/utils/global.dart';
@@ -26,9 +27,8 @@ class InitUtils {
       _setChainId();
       await NodeUtils.loadDbNodes();
 
-      await _openFavoriteTokensBox();
-      await _openNotificationsBox();
-      await _openRecipientBox();
+      // Initialize WalletConnect client
+      sl.get<WalletConnectService>().initClient();
     } catch (e) {
       rethrow;
     }
@@ -79,7 +79,9 @@ class InitUtils {
     zenon!.defaultKeyPair = kKeyStore!.getKeyPair(
       kDefaultAddressList.indexOf(kSelectedAddress),
     );
-
+    await _openFavoriteTokensBox();
+    await _openNotificationsBox();
+    await _openRecipientBox();
     await NodeUtils.initWebSocketClient();
     await _setWalletVersion();
     final baseAddress = await kKeyStore!.getKeyPair(0).address;
