+++ lib/widgets/main_app_container.dart
@@ -2,16 +2,13 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:app_links/app_links.dart';
-import 'package:badges/badges.dart' as badges;
 import 'package:clipboard_watcher/clipboard_watcher.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:flutter_vector_icons/flutter_vector_icons.dart';
-import 'package:hive/hive.dart';
 import 'package:logging/logging.dart';
-import 'package:lottie/lottie.dart';
 import 'package:provider/provider.dart';
 import 'package:wallet_connect_uri_validator/wallet_connect_uri_validator.dart';
 import 'package:window_manager/window_manager.dart';
@@ -19,7 +16,7 @@ import 'package:zenon_syrius_wallet_flutter/blocs/blocs.dart';
 import 'package:zenon_syrius_wallet_flutter/handlers/htlc_swaps_handler.dart';
 import 'package:zenon_syrius_wallet_flutter/main.dart';
 import 'package:zenon_syrius_wallet_flutter/model/model.dart';
-import 'package:zenon_syrius_wallet_flutter/services/i_web3wallet_service.dart';
+import 'package:zenon_syrius_wallet_flutter/services/wallet_connect_service.dart';
 import 'package:zenon_syrius_wallet_flutter/utils/app_colors.dart';
 import 'package:zenon_syrius_wallet_flutter/utils/clipboard_utils.dart';
 import 'package:zenon_syrius_wallet_flutter/utils/constants.dart';
@@ -46,8 +43,7 @@ enum Tabs {
   plasma,
   tokens,
   p2pSwap,
-  generation,
-  sync,
+  resyncWallet,
   accelerator,
   walletConnect,
 }
@@ -70,24 +66,33 @@ class _MainAppContainerState extends State<MainAppContainer>
     with TickerProviderStateMixin, ClipboardListener, WindowListener {
   late AnimationController _animationController;
   late Animation _animation;
-  late StreamSubscription _lockBlockStreamSubscription;
-  late StreamSubscription _incomingLinkSubscription;
-  late LockBloc _lockBloc;
-
-  Timer? _navigateToLockTimer;
-  TabController? _tabController;
-  TransferTabChild? _transferTabChild;
-  bool _initialUriIsHandled = false;
 
   final NodeSyncStatusBloc _netSyncStatusBloc = NodeSyncStatusBloc();
-  final _appLinks = AppLinks();
+
+  late StreamSubscription _lockBlockStreamSubscription;
+  late StreamSubscription _incomingLinkSubscription;
+
+  Timer? _navigateToLockTimer;
+
+  late LockBloc _lockBloc;
+
+  TabController? _tabController;
+
+  TransferTabChild? _transferTabChild;
+
   final FocusNode _focusNode = FocusNode(
     skipTraversal: true,
     canRequestFocus: false,
   );
 
+  bool _initialUriIsHandled = false;
+
+  final _appLinks = AppLinks();
+
   @override
   void initState() {
+    sl<WalletConnectService>().context = context;
+
     clipboardWatcher.addListener(this);
     windowManager.addListener(this);
 
@@ -107,7 +112,6 @@ class _MainAppContainerState extends State<MainAppContainer>
     _initLockBlock();
     _handleIncomingLinks();
     _handleInitialUri();
-
     super.initState();
   }
 
@@ -292,25 +296,14 @@ class _MainAppContainerState extends State<MainAppContainer>
     return <Tab>[
       if (kWcProjectId.isNotEmpty)
         Tab(
-          child: badges.Badge(
-            position: badges.BadgePosition.topEnd(top: -12.5, end: -12.5),
-            showBadge: (sl<IWeb3WalletService>().pairings.value.isNotEmpty),
-            badgeContent:
-                Text(sl<IWeb3WalletService>().pairings.value.length.toString()),
-            badgeStyle: const badges.BadgeStyle(
-              shape: badges.BadgeShape.circle,
-              badgeColor: AppColors.znnColor,
-              padding: EdgeInsets.all(3.5),
-            ),
-            child: SvgPicture.asset(
-              'assets/svg/walletconnect-logo.svg',
-              width: 24.0,
-              fit: BoxFit.fitWidth,
-              colorFilter: _isTabSelected(Tabs.walletConnect)
-                  ? const ColorFilter.mode(AppColors.znnColor, BlendMode.srcIn)
-                  : ColorFilter.mode(
-                      Theme.of(context).iconTheme.color!, BlendMode.srcIn),
-            ),
+          child: SvgPicture.asset(
+            'assets/svg/walletconnect-logo.svg',
+            width: 24.0,
+            fit: BoxFit.fitWidth,
+            colorFilter: _isTabSelected(Tabs.walletConnect)
+                ? const ColorFilter.mode(AppColors.znnColor, BlendMode.srcIn)
+                : ColorFilter.mode(
+                    Theme.of(context).iconTheme.color!, BlendMode.srcIn),
           ),
         ),
       Tab(
@@ -332,22 +325,12 @@ class _MainAppContainerState extends State<MainAppContainer>
         ),
       ),
       Tab(
-        child: badges.Badge(
-          position: badges.BadgePosition.topEnd(top: -10, end: -10),
-          showBadge: (Hive.box(kNotificationsBox).length > 0),
-          badgeContent: Text(Hive.box(kNotificationsBox).length.toString()),
-          badgeStyle: const badges.BadgeStyle(
-            shape: badges.BadgeShape.circle,
-            badgeColor: AppColors.znnColor,
-            padding: EdgeInsets.all(3.5),
-          ),
-          child: Icon(
-            Icons.notifications,
-            size: 24.0,
-            color: _isTabSelected(Tabs.notifications)
-                ? AppColors.znnColor
-                : Theme.of(context).iconTheme.color,
-          ),
+        child: Icon(
+          Icons.notifications,
+          size: 24.0,
+          color: _isTabSelected(Tabs.notifications)
+              ? AppColors.znnColor
+              : Theme.of(context).iconTheme.color,
         ),
       ),
       Tab(
@@ -360,10 +343,7 @@ class _MainAppContainerState extends State<MainAppContainer>
         ),
       ),
       Tab(
-        child: _getGenerationStatus(),
-      ),
-      Tab(
-        child: _getSyncStatus(),
+        child: _getPowGeneratingStatus(),
       ),
       Tab(
         child: _isTabSelected(Tabs.lock)
@@ -383,7 +363,7 @@ class _MainAppContainerState extends State<MainAppContainer>
     ];
   }
 
-  Widget _getSyncStatus() {
+  Widget _getWebsocketConnectionStatusStreamBuilder() {
     return StreamBuilder<SyncInfo>(
       stream: _netSyncStatusBloc.stream,
       builder: (_, snapshot) {
@@ -398,34 +378,8 @@ class _MainAppContainerState extends State<MainAppContainer>
     );
   }
 
-  Widget _getGenerationStatus() {
-    return StreamBuilder<PowStatus>(
-      stream: sl.get<PowGeneratingStatusBloc>().stream,
-      builder: (_, snapshot) {
-        if (snapshot.hasData && snapshot.data == PowStatus.generating) {
-          return Tooltip(
-            message: 'Generating Plasma',
-            child: Lottie.asset(
-              'assets/lottie/ic_anim_plasma_generation.json',
-              fit: BoxFit.contain,
-              width: 30.0,
-              repeat: true,
-            ),
-          );
-        }
-        return Tooltip(
-          message: 'Plasma generation idle',
-          child: Icon(
-            MaterialCommunityIcons.lightning_bolt,
-            color: Theme.of(context).iconTheme.color,
-          ),
-        );
-      },
-    );
-  }
-
   Widget _getSyncingStatusIcon(SyncState syncState, [SyncInfo? syncInfo]) {
-    String message = 'Connected and synced';
+    var message = 'Connected and synced';
 
     if (syncState != SyncState.notEnoughPeers &&
         syncState != SyncState.syncDone &&
@@ -436,13 +390,6 @@ class _MainAppContainerState extends State<MainAppContainer>
 
     if (syncState == SyncState.unknown) {
       message = 'Not ready';
-      return Tooltip(
-          message: message,
-          child: Icon(
-            Icons.sync_disabled,
-            size: 24.0,
-            color: _getSyncIconColor(syncState),
-          ));
     } else if (syncState == SyncState.syncing) {
       if (syncInfo != null) {
         if (syncInfo.targetHeight > 0 &&
@@ -450,44 +397,12 @@ class _MainAppContainerState extends State<MainAppContainer>
             (syncInfo.targetHeight - syncInfo.currentHeight) < 3) {
           message = 'Connected and synced';
           syncState = SyncState.syncDone;
-          return Tooltip(
-              message: message,
-              child: Lottie.asset(
-                'assets/lottie/ic_anim_live.json',
-                fit: BoxFit.contain,
-                width: 25.0,
-                repeat: true,
-              ));
-        } else if (syncInfo.targetHeight == 0 || syncInfo.currentHeight == 0) {
-          message = 'Started syncing with the network, please wait';
-          syncState = SyncState.syncing;
-          return Tooltip(
-              message: message,
-              child: Icon(Icons.sync,
-                  size: 24.0, color: _getSyncIconColor(syncState)));
         } else {
           message =
               'Sync progress: momentum ${syncInfo.currentHeight} of ${syncInfo.targetHeight}';
-          return Tooltip(
-            message: message,
-            child: SizedBox(
-              height: 25.0,
-              width: 25.0,
-              child: Center(
-                  child: CircularProgressIndicator(
-                backgroundColor: Theme.of(context).iconTheme.color,
-                color: _getSyncIconColor(syncState),
-                value: syncInfo.currentHeight / syncInfo.targetHeight,
-              )),
-            ),
-          );
         }
       } else {
         message = 'Syncing momentums';
-        return Tooltip(
-            message: message,
-            child: Icon(Icons.sync,
-                size: 24.0, color: _getSyncIconColor(syncState)));
       }
     } else if (syncState == SyncState.notEnoughPeers) {
       if (syncInfo != null) {
@@ -496,47 +411,17 @@ class _MainAppContainerState extends State<MainAppContainer>
             (syncInfo.targetHeight - syncInfo.currentHeight) < 20) {
           message = 'Connecting to peers';
           syncState = SyncState.syncing;
-          return Tooltip(
-              message: message,
-              child: SizedBox(
-                  height: 25.0,
-                  width: 25.0,
-                  child: Center(
-                      child: CircularProgressIndicator(
-                    backgroundColor: Theme.of(context).iconTheme.color,
-                    color: _getSyncIconColor(syncState),
-                    value: syncInfo.currentHeight / syncInfo.targetHeight,
-                  ))));
         } else if (syncInfo.targetHeight == 0 || syncInfo.currentHeight == 0) {
-          message = 'Connecting to peers, please wait';
+          message = 'Connecting to peers';
           syncState = SyncState.syncing;
-          return Tooltip(
-              message: message,
-              child: Icon(Icons.sync,
-                  size: 24.0, color: _getSyncIconColor(syncState)));
         } else {
           message =
               'Sync progress: momentum ${syncInfo.currentHeight} of ${syncInfo.targetHeight}';
           syncState = SyncState.syncing;
-          return Tooltip(
-              message: message,
-              child: SizedBox(
-                  height: 25.0,
-                  width: 25.0,
-                  child: Center(
-                      child: CircularProgressIndicator(
-                    backgroundColor: Theme.of(context).iconTheme.color,
-                    color: _getSyncIconColor(syncState),
-                    value: syncInfo.currentHeight / syncInfo.targetHeight,
-                  ))));
         }
       } else {
         message = 'Connecting to peers';
         syncState = SyncState.syncing;
-        return Tooltip(
-            message: message,
-            child: Icon(Icons.sync_problem,
-                size: 24.0, color: _getSyncIconColor(syncState)));
       }
     } else {
       message = 'Connected and synced';
@@ -544,13 +429,13 @@ class _MainAppContainerState extends State<MainAppContainer>
     }
 
     return Tooltip(
-        message: message,
-        child: Lottie.asset(
-          'assets/lottie/ic_anim_live.json',
-          fit: BoxFit.contain,
-          width: 25.0,
-          repeat: true,
-        ));
+      message: message,
+      child: Icon(
+        Icons.radio_button_unchecked,
+        size: 24.0,
+        color: _getSyncIconColor(syncState),
+      ),
+    );
   }
 
   Widget _getCurrentPageContainer() {
@@ -587,6 +472,7 @@ class _MainAppContainerState extends State<MainAppContainer>
         const NotificationsTabChild(),
         SettingsTabChild(
           _onChangeAutoLockTime,
+          _onResyncWalletPressed,
           onStepperNotificationSeeMorePressed: () => _navigateTo(
             Tabs.notifications,
           ),
@@ -594,7 +480,6 @@ class _MainAppContainerState extends State<MainAppContainer>
             Tabs.dashboard,
           ),
         ),
-        const SizedBox(),
         const SizedBox(),
         LockTabChild(_mainLockCallback, _afterAppInitCallback),
       ],
@@ -644,7 +529,6 @@ class _MainAppContainerState extends State<MainAppContainer>
     } else {
       _lockBloc.addEvent(LockEvent.navigateToDashboard);
     }
-
     _listenToAutoReceiveTxWorkerNotifications();
   }
 
@@ -652,6 +536,28 @@ class _MainAppContainerState extends State<MainAppContainer>
     sl<AutoReceiveTxWorker>().stream.listen((event) {
       sl<NotificationsBloc>().addNotification(event);
     });
+  }
+
+  void _onResyncWalletPressed() {
+    _navigateTo(Tabs.resyncWallet);
+  }
+
+  Widget _getPowGeneratingStatus() {
+    return StreamBuilder<PowStatus>(
+      stream: sl.get<PowGeneratingStatusBloc>().stream,
+      builder: (_, snapshot) {
+        if (snapshot.hasData && snapshot.data == PowStatus.generating) {
+          return const Tooltip(
+            message: 'Generating Plasma',
+            child: SyriusLoadingWidget(
+              size: 20.0,
+              strokeWidth: 2.5,
+            ),
+          );
+        }
+        return _getWebsocketConnectionStatusStreamBuilder();
+      },
+    );
   }
 
   bool _isTabSelected(Tabs page) =>
