+++ lib/widgets/modular_widgets/transfer_widgets/transfer_widgets.dart
@@ -1,5 +1,4 @@
-export 'latest_transactions/latest_transactions.dart';
-export 'pending_transactions/pending_transactions.dart';
+export 'latest_transactions/latest_transactions_transfer_widget.dart';
 export 'receive/receive_large.dart';
 export 'receive/receive_medium.dart';
 export 'receive/receive_small.dart';
