+++ lib/widgets/reusable_widgets/chart/standard_line_chart_bar_data.dart
@@ -7,11 +7,12 @@ class StandardLineChartBarData extends LineChartBarData {
     required Color color,
     required List<FlSpot>? spots,
   }) : super(
+          spots: spots,
           isCurved: false,
           color: color,
           barWidth: 3.0,
           isStrokeCapRound: true,
-          dotData: const FlDotData(
+          dotData: FlDotData(
             show: true,
           ),
           belowBarData: BarAreaData(
