+++ lib/widgets/reusable_widgets/infinite_scroll_table.dart
@@ -172,7 +172,6 @@ class _InfiniteScrollTableState<T> extends State<InfiniteScrollTable<T>> {
 
   @override
   void dispose() {
-    _scrollController.dispose();
     _pagingController.dispose();
     _blocListingStateSubscription.cancel();
     if (widget.disposeBloc) {
