+++ lib/widgets/tab_children_widgets/transfer_tab_child.dart
@@ -26,14 +26,9 @@ class _TransferTabChildState extends State<TransferTabChild> {
         _getSendCard(),
         _getReceiveCard(),
         const FluidCell(
+          width: kStaggeredNumOfColumns,
           child: LatestTransactions(),
-          width: kStaggeredNumOfColumns,
-          height: kStaggeredNumOfColumns / 3,
-        ),
-        const FluidCell(
-          child: PendingTransactions(),
-          width: kStaggeredNumOfColumns,
-          height: kStaggeredNumOfColumns / 3,
+          height: kStaggeredNumOfColumns / 2,
         ),
       ],
     );
