+++ pubspec.yaml
@@ -9,9 +9,9 @@ dependencies:
   flutter:
     sdk: flutter
 
-  fl_chart: ^0.63.0
+  fl_chart: ^0.62.0
   overlay_support: ^2.1.0
-  lottie: ^2.6.0
+  lottie: ^2.1.0
   dotted_border: ^2.0.0+2
   flutter_svg: ^2.0.0+1
   flutter_vector_icons: ^2.0.0
@@ -21,20 +21,20 @@ dependencies:
   flutter_staggered_grid_view: ^0.4.0
   flip_card: ^0.7.0
   auto_size_text: ^3.0.0
-  window_manager: ^0.3.5
-  stacked: ^3.4.1
-  path_provider: ^2.1.0
+  window_manager: ^0.3.4
+  stacked: ^3.1.2
+  path_provider: ^2.0.8
   rxdart: ^0.27.3
   hive: ^2.0.5
-  url_launcher: ^6.1.12
+  url_launcher: ^6.0.18
   provider: ^6.0.2
   intl: ^0.18.0
-  package_info_plus: ^4.1.0
-  device_info_plus: ^9.0.3
+  package_info_plus: ^4.0.2
+  device_info_plus: ^8.0.0
   infinite_scroll_pagination: ^3.1.0
-  share_plus: ^7.1.0
+  share_plus: ^7.0.2
   page_transition: ^2.0.4
-  file_selector: ^1.0.0
+  file_selector: ^0.9.2+2
   pretty_qr_code: ^2.0.3
   screenshot: ^2.1.0
   desktop_drop: ^0.4.0
@@ -47,7 +47,7 @@ dependencies:
       ref: d960c0bdc6dc553eaa75d5ddddb34cab530578b8
   json_rpc_2: ^3.0.2
   path: ^1.8.2
-  ffi: ^2.1.0
+  ffi: ^2.0.1
   get_it: ^7.2.0
   hex: ^0.2.0
   local_notifier: ^0.1.5
@@ -57,21 +57,20 @@ dependencies:
   logging: ^1.2.0
   collection: ^1.17.0
   app_links: ^3.4.3
-  walletconnect_flutter_v2: ^2.1.0
+  walletconnect_flutter_v2: ^2.0.14
   preference_list: ^0.0.1
-  screen_capturer: ^0.1.6
+  screen_capturer: ^0.1.2
+  uni_ocr: ^0.1.0
   zxing2: ^0.2.0
   image: ^4.0.10
   clipboard_watcher: ^0.2.0
   wallet_connect_uri_validator: ^0.1.0
   big_decimal: ^0.5.0
   ai_barcode_scanner: ^0.0.7
-  badges: ^3.1.1
-
 dev_dependencies:
-  build_runner: ^2.4.6
+  build_runner: ^2.1.7
   hive_generator: ^2.0.0
-  flutter_lints: ^2.0.2
+  flutter_lints: ^2.0.1
   dependency_validator: ^3.0.0
 
 flutter:
@@ -83,3 +82,8 @@ flutter:
     - assets/i18n/
     - assets/lottie/
     - assets/svg/
+
+# TODO: Are these really needed?
+dependency_overrides:
+    stream_channel: ^2.1.2
+    collection: ^1.17.2
